refactor(web): extract Judge cluster to pages/runView/JudgePanel.tsx (#1007 M2)
Pure motion of the Judge cluster (JUDGE_STAT_K..TriageMeter) verbatim,
including the W10 judge-verdict setInterval poll inside JudgePanel,
unchanged. RunView.tsx re-exports JudgePanel/JUDGE_POLL_MAX_TRIES/
TriageSummary so Judge.tsx and tests stay byte-identical.
